server/requirements.txt
fastapi==0.111.1
psycopg2==2.9.9
pydantic==2.6.4
python-dotenv==1.0.1
uvicorn==0.29.0
loguru==0.7.2
boto3
pytest
httpx